load products from backend
