+++ scripts/checkout.js
@@ -1,5 +1,9 @@
 import { renderOrderSummary } from "./checkout/orderSummary.js";
 import { renderPaymentSummary } from "./checkout/paymentSummary.js";
+import { loadDataFromBackend } from "../data/fromBackend.js";
 
-renderOrderSummary();
-renderPaymentSummary();
+loadDataFromBackend(function () {
+  renderOrderSummary();
+  renderPaymentSummary();
+})
+
