+++ scripts/checkout/paymentSummary.js
@@ -1,5 +1,5 @@
 import { cart } from "../../data/cart.js";
-import { products, getProduct } from "../../data/products.js";
+import { getProduct } from "../../data/fromBackend.js";
 import { getDeliveryOption } from "../../data/deliveryOptions.js";
 import { formatCurrency } from "../utils/money.js";
